requirements.txt
# Core ML Framework
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# Transformers and Hugging Face ecosystem
transformers==4.35.2
tokenizers==0.15.0
accelerate==0.24.1
datasets==2.14.7
evaluate==0.4.1

# Fine-tuning and Parameter Efficient Training
peft==0.6.2
bitsandbytes==0.41.3.post2
scipy==1.11.4

# Axolotl - QLoRA fine-tuning framework
axolotl>=0.9.0

# Quantization support
optimum==1.16.1

# API and Deployment
fastapi==0.104.1
uvicorn[standard]==0.24.0
gradio==4.7.1
pydantic==2.5.0

# Data processing and utilities
numpy==1.24.4
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
pillow==10.1.0
requests==2.31.0
tqdm==4.66.1

# Configuration and logging
pyyaml==6.0.1
wandb==0.16.1
tensorboard==2.15.1

# Development and code quality
ruff==0.1.7
black==23.11.0
pre-commit==3.6.0
pytest==7.4.3

# Jupyter and notebook support
jupyter==1.0.0
jupyterlab==4.0.9
ipykernel==6.27.1
ipywidgets==8.1.1

# System utilities
psutil==5.9.6
packaging==23.2

# Optional: Additional ML utilities
scikit-learn==1.3.2
safetensors==0.4.1